dom/tests/mochitest/geolocation/test_manyCurrentConcurrent.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=482260
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=482260">Mozilla Bug 482260</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
resume_geolocationProvider();
force_prompt(true);

var successCallbackCalled = 0;
var failureCallbackCalled = 0;
const totalResults = 100;

function successCallback(position) {
  successCallbackCalled++;
  testPassed();
}

function failureCallback(code) {
  failureCallbackCalled++;
  testPassed();
}

for (var x = 0; x < totalResults; x++)
  navigator.geolocation.getCurrentPosition(successCallback, failureCallback);

function testPassed() {
  if (successCallbackCalled + failureCallbackCalled != totalResults)
    return;
  is(failureCallbackCalled, 0, "no failure callbacks should have been received");
  reset_prompt();
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
